requirements_test.txt
pytest>=7.4.3
pytest-cov
pytest-homeassistant-custom-component>=0.13.88
pytest-asyncio
pytest-unordered
pylint
homeassistant-stubs
black
isort